Rearranged setup for better performance
For reasons unbeknownst to me, it was crashing if the NIST sync was before the "begin" function.  I didn't figure out why, but rearranging fixed it.  The most frustrating thing is that if I turned on the extra print statements for debugging... it magically worked every time, but as soon as they were off again.. it crashed.