Ensure all sensors return -9999 as bad result, added result verification
Now all variables should consistently return -9999 as a "bad" result.

Bad results will *not* be added to the values array, so only good results will be averaged.  This may mean that the actual number of results averaged is less than the number of results requested because bad values are tossed before averaging!