Allow >1 maxbotix, if different trigger pins
To create more than one with trigger pins **unattached** set the trigger pins as two different negative values.

Two sensors on the same stream with different trigger pins *should* work.

Two sensors on the same stream with the pulse of one acting as the trigger for the next will *NOT* work.