Removed extra overrides for Keller
The "overrride" functions were only doing exactly the base function, so there was no reason for an override.

I also moved the macrovariables for the variable numbers for the Keller into the KellerParent h, since they should be used in the KellerParent.cpp

Fixed the library.json.  The "version" tag must be the url of the github repo/web address unless the library is a registered library.  The "url," "library id," and "note" fields in the "dependencies" section are actually ignored by PlatformIO.  I just put them there as decoration.