Corrected stabilization timing in var array update
Previously was not properly waiting for a sensor to stabilize.  Measurements were only started if the sensor was stable, but if the sensor was not measuring it was assumed that it had failed, even if the reason it was not measuring was the unfinished stabilization time.

Had not noticed previously because most sensors have short stabilization times.