Added -9999 failure notification for Paleo terra Redox
NOTE: To verify the presance of absance of the sensor, it the code checks the result of the conversion, the i2c status bit, and the value of the config register. If the conversion result is 0, the i2c status 0, and the config is invalid, this corresponds to a detached sensor and the result will return -9999. This solution has the following known errata, if the sensor reads a value of 0 while operating in 12 bit mode (so config = 0x00) then it will return an erronious -9999, this is extremely unlikely (and we are currently operating in 18 bit mode) but should be documented.