Removed "checkForUpdate" function, forcing variable to specify if update
Decided that instead of having a magically "2 minute" age allowed for a variable's current value before the variable asks the sensor to update, to create a flag in the getValue function that specifies whether to ask the parent sensor to update or not.  By default, it will not update.  If you want the value to be updated by the sensor before the result is returned, you must specify getValue(true).