Maybe not internal client
Current issue: The internet connected loggers want to have an attached modem, and thus the modem.h is read in and pre-compiled with the .h files for those loggers.  But the logger modem itself isn't fully defined until compile time (pre-processor macros decided based on particular modem) so the final version of the logger modem instance is different from that of the pre-compiled instance.  Ugh.