Added power up to set up
Realized that this is necessary because the pin mode for the power pin isn't set to output until during the set-up function.  This means that running the powerUp function before running the setup function might not actually power the sensor.  It's probably always worked for us until now because we've always used the same pin for power and it's always been set as output.

NOTE;  The MODEM is still not powered in its set-up function.  I need to think more about the implications of this.