Do not need to wait for measurement completion within ADS1115 sensors
Not sure why I thought we needed to wait the measurement time before asking the ADC for a value.  We only have to wait the conversion delay between starting the ADC and getting a value from it and that (8ms) wait is included in the readADC_SingleEnded_V function.  I think it's safe to assume that the sensor will output a voltage whether or not the ADC is attached so that wait belongs in the same places it does for other sensors (ie, not here).