NOT unsetting timestamps or status bits for null sleep and powerdown fxns
It doesn't quite make sense to un-set the bits if nothing actually happened.  They should be left as-is.  In cases where the sleep function did nothing, bu the sensor was then powered down, the power down function un-sets the wake bits.  This means overrides aren't necessary for the clock and processor.