Update and rename DRWI_LTE.ino to DRWI_DigiLTE.ino
Changed filename to include Digi, to help differentiate this example from sim7080_LTE example.  Changed serial baud rate to 57600 to help with occasional missed characters.  Updated UUID code to match language currently on MMW.  Added line for setting pin 20 as output to help with proper operation of Mayfly1.0 boards (is okay to leave in code for Mayfly_v0.5 boards because pin 20 was unused unless something was using it through the 2x10 digital header).