LoggerBase: always power down modem after RTC sync
The test for "15 seconds before the next logging interval" has been wrong
for years, possibly since this code was written, with no apparent
consequence. The behavior is additionally confusing to users deploying the
devices and causes problems with logging as the modem won't get turned
off for a long time.

Remove it completely to solve the problems.