dataPublisherBase: widen sendEveryX and ditch sendOffset
sendEveryX will be used in the future for data buffering functionality. Increase
its width to an int to allow larger buffers when desired.

Delete sendOffset completely as there is little reason for that particular
functionality. The offset will be in effect set randomly using the time the
datalogger initially powers on.